Alteration to the Exame class and improving Main

+++ Nota.cs
@@ -1,11 +1,11 @@
 public class Nota {
 
 
-public int notaexame {get;set;}
+public float notaexame {get;set;}
 
 public Nota () {}
 
-public Nota (int notaexame) {
+public Nota (float notaexame) {
     this.notaexame = notaexame;
 }
